voice-agent/pyproject.toml
[project]
name = "voice-agent"
version = "0.1.0"
description = ""
authors = [
    {name = "anujsb",email = "anujsbhuyar@gmail.com"}
]
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.135.1,<0.136.0)",
    "uvicorn (>=0.42.0,<0.43.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "twilio (>=9.10.3,<10.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
